requirements.txt
pytorch-lightning
wandb
torchinfo